test: refactor: split tests to different files
test/index.js: entrypoint; runs all tests like previously
test/detector.test.js: unit tests for PhishingDetector implementation
test/config.test.js: validation tests for src/config.json
test/test.util.js: test utility functions
